+++ src/main/python/tranquilitybase/gcpdac_mock/core/application.py
@@ -1,25 +1,22 @@
-# Supports all actions concerning sandboxes
-import json
-import logging
 import os
+import logging
+import json
 import random
 from pprint import pformat
 
-from config import counter
+from src.main.python.tranquilitybase.gcpdac_mock.config import counter
 
-logger = logging.getLogger("sandbox")
+logger = logging.getLogger("application")
 
-sandbox_response_json = ""
-sandbox_response_file = "gcpdac/sandbox_response_example.json"
-with open(sandbox_response_file, "r") as fh:
-    sandbox_response_json = json.load(fh)
-
+application_response_json = ""
+application_response_file = "core/sample_responses/application/application_response_example.json"
+with open(application_response_file, "r") as fh:
+    application_response_json = json.load(fh)
 
 def next_taskid():
     task_cnt = next(counter)
     taskid = f"MOCKTASKID{task_cnt}"
     return taskid
-
 
 def get_random_status():
     r = random.randint(0, 10)
@@ -39,9 +36,13 @@ def get_status():
         return "SUCCESS"
 
 
-def create_async(sandboxDetails):
-    logger.debug(pformat(sandboxDetails))
-    taskid = next_taskid()
+def create_async(applicationDetails):
+    """
+    Return just the task_id.
+    """
+
+    logger.debug(pformat(applicationDetails))
+    taskid =  next_taskid()
     logger.info("Task ID %s", taskid)
     context = {
         "taskid": f"{taskid}"
@@ -50,8 +51,11 @@ def create_async(sandboxDetails):
 
 
 def delete_async(oid):
+    """
+    Return just the task_id.
+    """
     logger.debug("Id is {}".format(oid))
-    taskid = next_taskid()
+    taskid =  next_taskid()
     logger.info("Task ID %s", taskid)
     context = {
         "taskid": f"{taskid}"
@@ -59,23 +63,23 @@ def delete_async(oid):
     return context, 200
 
 
-def create_sandbox_result(taskid):
-    logger.info("CREATE sandbox RESULT %s", format(taskid))
+def create_application_result(taskid):
+    logger.info("CREATE application RESULT %s",format(taskid))
 
     retval = {
         "status": get_status()
     }
     if retval.get('status') == "SUCCESS" or retval.get('status') == "FAILURE":
-        retval["payload"] = json.dumps(sandbox_response_json['payload'])
+        retval["payload"] = json.dumps(application_response_json['payload'])
     return retval, 201
 
 
-def delete_sandbox_result(taskid):
-    logger.info("DELETE sandbox RESULT %s", format(taskid))
+def delete_application_result(taskid):
+    logger.info("DELETE application RESULT %s",format(taskid))
 
     retval = {
         "status": get_status()
     }
     if retval.get('status') == "SUCCESS" or retval.get('status') == "FAILURE":
-        retval["payload"] = json.dumps(sandbox_response_json['payload'])
+        retval["payload"] = json.dumps(application_response_json['payload'])
     return retval, 200
