+++ src/main/python/tranquilitybase/gcpdac_mock/core/sample_responses/application/application_response_example.json
