+++ src/main/python/tranquilitybase/gcpdac_mock/core/sample_responses/folder/folder_response_example.json
