+++ src/main/python/tranquilitybase/gcpdac_mock/core/sample_responses/solution/solution_response_example.json
