--- src/main/python/tranquilitybase/gcpdac_mock/mock_gcp_dac/README.md
@@ -1,7 +1,0 @@
-# Useful commands
-
-## Build a mock container image using the Dockerfile used by the CICD workflow
-docker build -t gcr.io/eagle-console-resources/tb-gcp-dac:mock .
-docker push gcr.io/eagle-console-resources/tb-gcp-dac:mock
-
-## References
