--- src/main/python/tranquilitybase/gcpdac_mock/mock_gcp_dac/gcpdac/extendedSchemas.py
@@ -1,47 +1,0 @@
-from marshmallow import Schema, fields
-
-
-class HealthSchema(Schema):
-    """
-    Schema for health check.
-    """
-    def __init__(self, **kwargs):
-        """
-        Constructor for this class.
-        """
-        super().__init__(**kwargs)
-
-    status = fields.Str()
-
-
-class ResponseSchema(Schema):
-    """
-    Schema for HS Response.
-    """
-    def __init__(self, **kwargs):
-        """
-        Constructor for this class.
-        """
-        super().__init__(**kwargs)
-
-    id = fields.Int()
-    deployed = fields.Bool()
-    deploymentState = fields.Str()
-    statusId = fields.Int()
-    statusCode = fields.Str()
-    statusMessage = fields.Str()
-    taskId = fields.Int()
-
-
-class MetadataSchema(Schema):
-    """
-    Schema for Metadata Response.
-    """
-    def __init__(self, **kwargs):
-        """
-        Constructor for this class.
-        """
-        super().__init__(**kwargs)
-
-    root_folder_id = fields.Str()
-    shared_vpc_host_project = fields.Str()
